Add an `asini add` command
This adds and installs dependencies across packages in the repo.

Can be limited to certain packages with `--scope`.

Default version is latest with caret range.
