+++ test/fixtures/AddCommand/basic/asini.json
@@ -1,0 +1,4 @@
+{
+  "asini": "__TEST_VERSION__",
+  "version": "1.0.0"
+}
