+++ test/fixtures/AddCommand/basic/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "independent"
+}
