+++ test/fixtures/AddCommand/basic/packages/package-4/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "package-4",
+  "version": "1.0.0"
+}
